requirements.txt
google-generativeai>=0.3.0
triton>=2.0.0
numpy>=1.24.0
torch>=2.0.0
